adding optional playbutton param, to show a play button
having the song play after page load was a little annoying for me when I
used the embeddable option.  This commit introduces a new &playbutton
param.  If present a link labelled '▸ Click to play' is appended and the
tune will not start until you click it.

A new global object songOfGitHub was introduced to deal with this, and
allWeeks array was pulled up.